+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/jpa/SequenceInstanceCountByType.java
@@ -1,0 +1,4 @@
+package ch.admin.bit.jeap.messaging.sequentialinbox.jpa;
+
+record SequenceInstanceCountByType(String sequenceType, long count) {
+}